+++ example/lib/main.dart
@@ -1,70 +1,43 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart' as DotEnv;
+import 'package:flutter/services.dart';
+import 'package:flutter_dotenv/flutter_dotenv.dart' as dotenv;
 
 Future main() async {
-  await DotEnv.load();
+  await dotenv.load(mergeWith: {
+    'TEST_VAR': '5',
+  }); // mergeWith optional, you can include Platform.environment for Mobile/Desktop app
+
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) => MaterialApp(
-        title: 'Flutter Demo',
-        theme: ThemeData(
-          primarySwatch: Colors.blue,
-          visualDensity: VisualDensity.adaptivePlatformDensity,
-        ),
-        home: MyHomePage(title: 'Flutter Demo Home Page'),
-      );
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    print(DotEnv.env);
-    setState(() {
-      _counter++;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(
-          title: Text(widget.title),
-        ),
-        body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Text(
-                'You have pushed the button this many times:}',
+        title: 'Dotenv Demo',
+        home: Scaffold(
+          appBar: AppBar(
+            title: Text('Dotenv Demo'),
+          ),
+          body: SingleChildScrollView(
+            child: FutureBuilder<String>(
+              future: rootBundle.loadString('assets/.env'),
+              initialData: '',
+              builder: (context, snapshot) => Container(
+                padding: EdgeInsets.all(50),
+                child: Column(
+                  children: [
+                    Text(
+                      'Env map: ${dotenv.env.toString()}',
+                    ),
+                    Divider(thickness: 5),
+                    Text('Original'),
+                    Divider(),
+                    Text(snapshot.data ?? ''),
+                  ],
+                ),
               ),
-              Text(
-                '$_counter',
-                style: Theme.of(context).textTheme.headline4,
-              ),
-              Text(
-                'Env map: ${DotEnv.env}',
-              ),
-            ],
+            ),
           ),
         ),
-        floatingActionButton: FloatingActionButton(
-          onPressed: _incrementCounter,
-          tooltip: 'Increment',
-          child: Icon(Icons.add),
-        ), // This trailing comma makes auto-formatting nicer for build methods.
       );
 }