+++ lib/src/dotenv.dart
@@ -58,11 +58,25 @@ Future load(
   _isInitialized = true;
 }
 
+Future testLoad(
+    {String fileInput = '',
+    Parser parser = const Parser(),
+    Map<String, String> mergeWith = const {}}) async {
+  clean();
+  final linesFromFile = fileInput.split('\n');
+  final linesFromMergeWith =
+      mergeWith.entries.map((entry) => "${entry.key}=${entry.value}").toList();
+  final allLines = linesFromMergeWith..addAll(linesFromFile);
+  final envEntries = parser.parse(allLines);
+  _envMap.addAll(envEntries);
+  _isInitialized = true;
+}
+
 /// True if all supplied variables have nonempty value; false otherwise.
 /// Differs from [containsKey](dart:core) by excluding null values.
 /// Note [load] should be called first.
 bool isEveryDefined(Iterable<String> vars) =>
-    vars.every((k) => _envMap[k] != null && _envMap[k].isNotEmpty);
+    vars.every((k) => _envMap[k]?.isNotEmpty ?? false);
 
 Future<List<String>> _getEntriesFromFile(String filename) async {
   try {