+++ CHANGELOG.md
@@ -9,6 +9,19 @@ Release notes are available on [github][notes].
 [pub-semver-readme]: https://pub.dartlang.org/packages/pub_semver
 [notes]: https://github.com/java-james/flutter_dotenv/releases
 
+4.0.0-nullsafety.0
+-----
+
+- [BREAKING] Opt into null-safety
+- [deps] Upgrade dart sdk constraints to ```>=2.12.0-0 <3.0.0```
+- [new] Allow for escape of $ ' " and \n characters
+- [fix] Ensure swallow function only removes leading 'export' keyword
+- [fix] Retain spaces within single or double quotes
+- [fix] Allow for comments after matching end quotes
+- [new] Migrate to null safety
+- [new] Create unit test cases for parse
+
+
 #### 3.1.0
 
 - [new] Allow merging with a custom map on load